Create SideModal component
